Implement incidentTimelineService.deleteBy

+++ backend/backend/services/incidentService.js
@@ -181,9 +181,11 @@ module.exports = {
                     const { _id } = monitorStatus;
                     await MonitorStatusService.deleteBy({ _id }, userId);
                 }
-                const incidentTimeline = await IncidentTimelineService.findBy({_id: incident._id});
-                for(const event of incidentTimeline){
-                    await IncidentTimelineService.deleteBy({_id:event._id},)
+                const incidentTimeline = await IncidentTimelineService.findBy({
+                    incidentId: incident._id,
+                });
+                for (const event of incidentTimeline) {
+                    await IncidentTimelineService.deleteBy({ _id: event._id }, userId);
                 }
             }
             return incident;
